rename the delete_tag to just delete
