+++ tests/commands/tags/test_rename_tag.py
@@ -7,7 +7,7 @@ from gitblend.commands.tags.rename import run
 
 class TestRenameTagCommand(unittest.TestCase):
 
-    @patch("gitblend.commands.tags.delete_tag.run")
+    @patch("gitblend.commands.tags.delete.run")
     @patch("gitblend.commands.tags.create.run")
     def test_rename_tag_success(self, mock_create_tag, mock_delete_tag):
         args = MagicMock(old_tag="v1.0", new_tag="v1.1")
@@ -22,7 +22,7 @@ class TestRenameTagCommand(unittest.TestCase):
                 "✅ Tag 'v1.0' has been renamed to 'v1.1' successfully."
             )
 
-    @patch("gitblend.commands.tags.delete_tag.run")
+    @patch("gitblend.commands.tags.delete.run")
     @patch("gitblend.commands.tags.create.run")
     def test_rename_tag_delete_failure(self, mock_create_tag, mock_delete_tag):
         mock_delete_tag.side_effect = SystemExit(1)
@@ -39,7 +39,7 @@ class TestRenameTagCommand(unittest.TestCase):
                 "❌ Error while renaming tag 'v1.0' to 'v1.1': 1", file=sys.stderr
             )
 
-    @patch("gitblend.commands.tags.delete_tag.run")
+    @patch("gitblend.commands.tags.delete.run")
     @patch("gitblend.commands.tags.create.run")
     def test_rename_tag_create_failure(self, mock_create_tag, mock_delete_tag):
         mock_create_tag.side_effect = SystemExit(1)
